Extract useActions to root
