+++ packages/app/src/features/actions/flavours/approve/logic/approve-action.ts
@@ -1,34 +1,10 @@
-import { ensureConfigTypes, useWrite } from '@/domain/hooks/useWrite'
-import { MarketInfo } from '@/domain/market-info/marketInfo'
+import { ensureConfigTypes } from '@/domain/hooks/useWrite'
 import { allowance, allowanceQueryKey } from '@/domain/market-operations/allowance/query'
 import { normalizeErc20AbiForToken } from '@/domain/market-operations/normalizeErc20Abi'
-import { TokensInfo } from '@/domain/wallet/useTokens/TokenInfo'
+import { ActionConfig, ActionContext } from '@/features/actions/logic/types'
 import { toBigInt } from '@/utils/bigNumber'
-import { QueryKey, UseQueryOptions, UseQueryResult, queryOptions } from '@tanstack/react-query'
-import { Address } from 'viem'
-import { Config } from 'wagmi'
+import { queryOptions } from '@tanstack/react-query'
 import { ApproveAction } from '../types'
-
-export interface ActionContext {
-  marketInfo?: MarketInfo
-  tokensInfo?: TokensInfo
-  wagmiConfig: Config
-  account: Address
-  chainId: number
-}
-
-type InitialParamsQueryOptions = UseQueryOptions<any, Error, { canBeSkipped: boolean }, QueryKey>
-export type InitialParamsQueryResult = UseQueryResult<{ canBeSkipped: boolean }>
-type VerifyTransactionQueryOptions = UseQueryOptions<any, Error, { success: boolean }, QueryKey>
-export type VerifyTransactionResult = UseQueryResult<{ success: boolean }>
-
-export interface ActionConfig {
-  initialParamsQueryOptions: () => InitialParamsQueryOptions
-  getWriteConfig: (initialParams: InitialParamsQueryResult) => Parameters<typeof useWrite>[0]
-  verifyTransactionQueryOptions: () => VerifyTransactionQueryOptions
-  invalidates: () => QueryKey[]
-  beforeWriteCheck?: () => {}
-}
 
 export interface CreateApproveActionConfigParams {
   action: ApproveAction
